+++ provider/pkg/provider/util.go
@@ -5,6 +5,9 @@ package provider
 import (
 	"encoding/json"
 	"fmt"
+	"os"
+	"os/user"
+	"path/filepath"
 	"reflect"
 	"strings"
 
@@ -79,6 +82,44 @@ func fqName(namespace, name string) string {
 // Kubeconfig helpers.
 // --------------------------------------------------------------------------
 
+// parseKubeconfigString takes a string that contains either a path to a kubeconfig file
+// or the contents of a kubeconfig (YAML or JSON).
+func parseKubeconfigString(pathOrContents string) (*clientapi.Config, error) {
+	var contents string
+
+	if pathOrContents == "" {
+		return &clientapi.Config{}, nil
+	}
+
+	// Handle the '~' character if it is set in the config string. Normally, this would be expanded by the shell
+	// into the user's home directory, but we have to do that manually if it is set in a config value.
+	homeDir := func() string {
+		// Ignore errors. The filepath will be checked later, so we can handle failures there.
+		usr, _ := user.Current()
+		return usr.HomeDir
+	}
+	if pathOrContents == "~" {
+		// In case of "~", which won't be caught by the "else if"
+		pathOrContents = homeDir()
+	} else if strings.HasPrefix(pathOrContents, "~/") {
+		pathOrContents = filepath.Join(homeDir(), pathOrContents[2:])
+	}
+
+	// If the variable is a valid filepath, load the file and parse the contents as a k8s config.
+	_, err := os.Stat(pathOrContents)
+	if err == nil {
+		b, err := os.ReadFile(pathOrContents)
+		if err != nil {
+			return nil, err
+		}
+		contents = string(b)
+	} else { // Assume the contents are a k8s config.
+		contents = pathOrContents
+	}
+
+	return clientcmd.Load([]byte(contents))
+}
+
 // parseKubeconfigPropertyValue takes a PropertyValue that possibly contains a raw kubeconfig
 // (YAML or JSON) string or map and attempts to unmarshal it into a Config struct. If the property value
 // is empty, an empty Config is returned.
@@ -87,31 +128,31 @@ func parseKubeconfigPropertyValue(kubeconfig resource.PropertyValue) (*clientapi
 		return &clientapi.Config{}, nil
 	}
 
-	var cfg []byte
+	var cfg string
 	if kubeconfig.IsString() {
-		cfg = []byte(kubeconfig.StringValue())
+		cfg = kubeconfig.StringValue()
 	} else if kubeconfig.IsObject() {
 		raw := kubeconfig.ObjectValue().Mappable()
 		jsonBytes, err := json.Marshal(raw)
 		if err != nil {
 			return nil, fmt.Errorf("failed to unmarshal kubeconfig: %v", err)
 		}
-		cfg = jsonBytes
+		cfg = string(jsonBytes)
 	} else {
 		return nil, fmt.Errorf("unexpected kubeconfig format, type: %v", kubeconfig.TypeString())
 	}
-	config, err := clientcmd.Load(cfg)
+	config, err := parseKubeconfigString(cfg)
 	if err != nil {
 		return nil, fmt.Errorf("failed to parse kubeconfig: %v", err)
 	}
-
 	return config, nil
 }
 
 // getActiveClusterFromConfig gets the current cluster from a kubeconfig, accounting for provider overrides.
-func getActiveClusterFromConfig(config *clientapi.Config, overrides resource.PropertyMap) *clientapi.Cluster {
+// If the config is nil or the active cluster could not be found, false is returned.
+func getActiveClusterFromConfig(config *clientapi.Config, overrides resource.PropertyMap) (*clientapi.Cluster, bool) {
 	if config == nil || len(config.Clusters) == 0 {
-		return &clientapi.Cluster{}
+		return &clientapi.Cluster{}, false
 	}
 
 	currentContext := config.CurrentContext
@@ -121,7 +162,7 @@ func getActiveClusterFromConfig(config *clientapi.Config, overrides resource.Pro
 
 	activeContext := config.Contexts[currentContext]
 	if activeContext == nil {
-		return &clientapi.Cluster{}
+		return &clientapi.Cluster{}, false
 	}
 	activeClusterName := activeContext.Cluster
 
@@ -130,10 +171,10 @@ func getActiveClusterFromConfig(config *clientapi.Config, overrides resource.Pro
 		activeCluster = config.Clusters[val.StringValue()]
 	}
 	if activeCluster == nil {
-		return &clientapi.Cluster{}
+		return &clientapi.Cluster{}, false
 	}
 
-	return activeCluster
+	return activeCluster, true
 }
 
 // pruneMap builds a pruned map by recursively copying elements from the source map that have a matching key in the